+++ src/main/java/co/edu/corposucre/productionfood/usuario/dto/UsuarioResponse.java
@@ -1,11 +1,11 @@
 package co.edu.corposucre.productionfood.usuario.dto;
 
 public record UsuarioResponse(
-    Long idUsuario,
+    Integer idUsuario,
     String nombre,
     String correo,
     boolean activo,
     RolResponse rol
 ) {
-    public record RolResponse(Long idRol, String nombre) {}
+    public record RolResponse(Integer idRol, String nombre) {}
 }
